requirements.txt
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
types-beautifulsoup4==4.12.0.20241020
types-html5lib==1.1.11.20241018
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0

pytest~=8.3.4
pytest-order==1.3.0
allure-pytest==2.15.0
fuzz~=0.1.1
fuzzywuzzy~=0.18.0